refactor: Split util::insert into its own file
